skip testing on scanners

+++ experiments/run-exp.sh
@@ -52,29 +52,30 @@ else
 fi
 echo
 
-if [ -f ./models_table.txt ]; then
-    if [ "$2" != "y" ]; then
-        echo "./models_table.txt exists. Should I remove it? (y/n)"
-        read proceed
-    else
-        proceed="y"
+if grep -q "sampler_scanner: sampler" $CONFIG_FILE; then
+    if [ -f ./models_table.txt ]; then
+        if [ "$2" != "y" ]; then
+            echo "./models_table.txt exists. Should I remove it? (y/n)"
+            read proceed
+        else
+            proceed="y"
+        fi
+        if [ "$proceed" = "y" ]; then
+            rm -f all_models_table.txt models_table.txt
+            echo "./models_table.txt Removed."
+        fi
     fi
-    if [ "$proceed" = "y" ]; then
-        rm -f all_models_table.txt models_table.txt
-        echo "./models_table.txt Removed."
+    for filename in $( ls -rt models/model_*-v*.json ); do
+        echo $filename >> all_models_table.txt
+    done
+    awk 'NR == 1 || NR % 20 == 0' all_models_table.txt > models_table.txt
+
+    echo "Evaluating the models on the testing data..."
+    if ! $RUN_SPARROW test $CONFIG_FILE 2> $PREDICTION_LOG; then
+        echo "Evaluation failed."
+        cat $PREDICTION_LOG
+        exit
     fi
-fi
-for filename in $( ls -rt models/model_*-v*.json ); do
-    echo $filename >> all_models_table.txt
-done
-awk 'NR == 1 || NR % 20 == 0' all_models_table.txt > models_table.txt
-
-
-echo "Evaluating the models on the testing data..."
-if ! $RUN_SPARROW test $CONFIG_FILE 2> $PREDICTION_LOG; then
-    echo "Evaluation failed."
-    cat $PREDICTION_LOG
-    exit
 fi
 
 echo "All done."
